backend/requirements.txt
requests
beautifulsoup4
langchain
openai
chromadb
fastapi
uvicorn
python-dotenv
langchain-community
tiktoken